SPI RTC connect with BB board

+++ FunctionPeriph/inc/delay_systick.h
@@ -1,9 +1,9 @@
 #ifndef __DELAY_SYSYTICK_H
 #define __DELAY_SYSYTICK_H
 
-void delay_ms (u16 delay_time);
-void delay_s (u16 delay_time);
-void delay_us (u16 delay_time);
+void delay_ms (uint32_t delay_time);
+void delay_s (uint32_t delay_time);
+void delay_us (uint32_t delay_time);
 void delay_counter_down(void);
 
 #endif
